# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Linx API
  version: "1"
  description: >-
    The API used by the admin portal, the web client, the `linx` CLI and
    every integration (docs/API.md).
  license:
    name: Apache-2.0
    identifier: Apache-2.0
security:
  - bearer: []
paths:
...
  /api/v1/event-types:
    get:
      operationId: listEventTypes
      summary: Webhook and admin-alert event types this server can emit
      description: The event catalog webhook endpoints subscribe to (docs/API.md §4).
      tags: [meta]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of event types.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventTypeList"
        default:
          $ref: "#/components/responses/Problem"
...
  /api/v1/users/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getUser
      summary: Get a person
      tags: [people]
      security:
        - bearer: [users:read]
      responses:
        "200":
          description: The person.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateUser
      summary: Change a person
      description: >-
        JSON Merge Patch: only the fields sent change. `disabled: true` ends
        every session of theirs at once (docs/WEB.md §4).
      tags: [people]
      security:
        - bearer: [users:write]
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/UserPatch"
      responses:
        "200":
          description: The person as they now are.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: disableUser
      summary: Disable a person
      description: >-
        Stops them signing in and ends every session of theirs at once
        (docs/WEB.md §4). Disabling a disabled person succeeds and changes
        nothing. Never deletes their history.
      tags: [people]
      security:
        - bearer: [users:write]
      responses:
        "204":
          description: Disabled.
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: >-
        An API key (`linx_<id>_<secret>`) or an access token from
        /oauth/token. Operations list the scopes they need (docs/API.md §3).
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      description: Maximum items per page (docs/API.md §2).
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      description: Opaque pagination cursor from a previous page's `next_cursor`.
      schema:
        type: string
...
  responses:
    Problem:
      description: An error (docs/API.md §2).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    EventType:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          description: e.g. `call.missed`.
        description:
          type: string
    EventTypeList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EventType"
        next_cursor:
          type: string
          description: Pass as `?cursor=` to fetch the next page; absent on the last page.
    Role:
      type: string
      enum: [system_admin, admin, user, reporter]
...
    Problem:
      description: RFC 9457 `application/problem+json` (docs/API.md §2).
      type: object
      required: [type, title, status, code, detail]
      properties:
        type:
          type: string
          description: A URI identifying the problem type; `about:blank` when none is registered.
        title:
          type: string
          description: Short, human-readable summary of the status code.
        status:
          type: integer
        code:
          type: string
          description: Stable machine-readable error code, e.g. `scope_missing`.
        detail:
          type: string
          description: Plain-language explanation. Never a stack trace.
        instance:
          type: string
...
    User:
      type: object
      required: [id, email, name, role, mfa_enabled, disabled, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        extension_id:
          type: string
          format: uuid
        mfa_enabled:
          type: boolean
        disabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserPatch:
      description: JSON Merge Patch; fields not sent stay as they are.
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        role:
          $ref: "#/components/schemas/Role"
        extension_id:
          type: string
          format: uuid
        disabled:
          type: boolean
...